Уберет Header из Favorites
