+++ project/src/pages/favorites-empty/favorites-empty.tsx
@@ -1,4 +1,4 @@
-function FavoriteEmpty(): JSX.Element {
+function FavoritesEmpty(): JSX.Element {
   return (
     <main className="page__main page__main--favorites page__main--favorites-empty">
       <div className="page__favorites-container container">
@@ -14,4 +14,4 @@ function FavoriteEmpty(): JSX.Element {
   );
 }
 
-export default FavoriteEmpty;
+export default FavoritesEmpty;
